@@ -1,5 +1,5 @@
 class Category < ActiveRecord::Base
-  attr_accessible :name, :sequence
-  default_scope :order => 'sequence ASC'
+  # attr_accessible :name, :sequence
+  default_scope {order('sequence ASC') }
   
 end
