@@ -1,6 +1,6 @@
 class Stock < ActiveRecord::Base
-  attr_accessible :category_id, :name, :website
+  # attr_accessible :category_id, :name, :website
   belongs_to :category
-  default_scope :order => 'name ASC'
+  default_scope { order('name ASC') }
   
 end
